Make FileChooserUI extend abstract ChooserUI
